requirements.txt
cachetools==5.5.2
camelot-py==0.11.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
decorator==5.1.1
distro==1.9.0
dotenv==0.9.9
Elixir==0.7.1
et_xmlfile==2.0.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
idna==3.7
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pbr==6.1.0
pdfminer.six==20240706
proto-plus==1.26.1
protobuf==6.31.0
py4j==0.10.9.7
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMuPDF==1.24.13
pypdf==5.1.0
pyspark==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
rsa==4.9.1
six==1.16.0
SQLAlchemy==1.4.0
sqlalchemy-migrate==0.11.0
sqlparse==0.5.1
tabula-py==2.10.0
tabulate==0.9.0
Tempita==0.5.2
tzdata==2024.2
urllib3==2.2.2
xlrd==0.7.1
xlwt==0.7.2
google-cloud-secret-manager
pandas_gbq
